+++ src/display/output/mod.rs
@@ -46,7 +46,6 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-pub(crate) mod banner;
 pub(crate) mod formats;
 pub(crate) mod populate;
 pub(crate) mod quotes;